fix(conversation): show full GitHub thread with real author chip and verb
