+++ src/components/common/index.ts
@@ -2,3 +2,5 @@ export * from './SearchInput';
 export * from './linkBehavior';
 export * from './WatchlistButton';
 export * from './DataTable';
+export { default as ConversationTimeline } from './ConversationTimeline';
+export * from './ConversationTimeline';
