+++ src/components/prs/PRComments.tsx
@@ -1,40 +1,32 @@
-import React from 'react';
-import { formatDate } from '../../utils/format';
-import {
-  Box,
-  Typography,
-  Avatar,
-  Paper,
-  Link,
-  CircularProgress,
-  Chip,
-} from '@mui/material';
-import { alpha } from '@mui/material/styles';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import rehypeRaw from 'rehype-raw';
+import React, { useEffect, useMemo, useState } from 'react';
+import axios from 'axios';
+import { Box, Typography, CircularProgress } from '@mui/material';
 import { usePullRequestComments } from '../../api';
 import {
   type PullRequestComment,
   type PullRequestDetails,
 } from '../../api/models/Dashboard';
-import { STATUS_COLORS, UI_COLORS, scrollbarSx } from '../../theme';
-import 'github-markdown-css/github-markdown-dark.css'; // Import standard GitHub Dark styles
-
-/** A comment or the PR description rendered in the conversation timeline. */
-type ConversationItem = Omit<
-  PullRequestComment,
-  'id' | 'updatedAt' | 'htmlUrl'
-> & {
-  id: number | string;
-  isDescription?: boolean;
-};
+import { getGithubAvatarSrc } from '../../utils/ExplorerUtils';
+import {
+  ConversationTimeline,
+  EMPTY_BODY_PLACEHOLDER,
+  githubProfileUrl,
+  type ConversationItem,
+} from '../common';
 
 interface PRCommentsProps {
   repository: string;
   pullRequestNumber: number;
   prDetails: PullRequestDetails;
 }
+
+const commentToItem = (c: PullRequestComment): ConversationItem => ({
+  id: `comment-${c.id}`,
+  user: c.user,
+  body: c.body || EMPTY_BODY_PLACEHOLDER,
+  createdAt: c.createdAt,
+  authorAssociation: c.authorAssociation || 'NONE',
+});
 
 const PRComments: React.FC<PRCommentsProps> = ({
   repository,
@@ -46,6 +38,51 @@ const PRComments: React.FC<PRCommentsProps> = ({
     isLoading,
     error,
   } = usePullRequestComments(repository, pullRequestNumber);
+  const [bodyAssociation, setBodyAssociation] = useState<string>('NONE');
+
+  useEffect(() => {
+    setBodyAssociation('NONE');
+    if (!repository || !pullRequestNumber) return;
+    let cancelled = false;
+    axios
+      .get<{ author_association?: string | null }>(
+        `https://api.github.com/repos/${repository}/pulls/${pullRequestNumber}`,
+      )
+      .then((res) => {
+        if (cancelled) return;
+        setBodyAssociation(res.data.author_association || 'NONE');
+      })
+      .catch((err) => {
+        if (!cancelled) {
+          console.error('Failed to fetch PR author_association', err);
+        }
+      });
+    return () => {
+      cancelled = true;
+    };
+  }, [repository, pullRequestNumber]);
+
+  const items = useMemo<ConversationItem[]>(() => {
+    const body: ConversationItem = {
+      id: 'pr-description',
+      user: {
+        login: prDetails.authorLogin,
+        avatarUrl: getGithubAvatarSrc(prDetails.authorLogin),
+        htmlUrl: githubProfileUrl(prDetails.authorLogin),
+      },
+      body: prDetails.description || EMPTY_BODY_PLACEHOLDER,
+      createdAt: prDetails.createdAt,
+      authorAssociation: bodyAssociation,
+      isDescription: true,
+    };
+    return [body, ...(comments ?? []).map(commentToItem)];
+  }, [
+    prDetails.authorLogin,
+    prDetails.description,
+    prDetails.createdAt,
+    bodyAssociation,
+    comments,
+  ]);
 
   if (isLoading) {
     return (
@@ -65,327 +102,7 @@ const PRComments: React.FC<PRCommentsProps> = ({
     );
   }
 
-  const allItems: ConversationItem[] = [
-    {
-      id: 'pr-description',
-      user: {
-        login: prDetails.authorLogin,
-        avatarUrl: `https://avatars.githubusercontent.com/${prDetails.authorLogin}`,
-        htmlUrl: `https://github.com/${prDetails.authorLogin}`,
-      },
-      body: prDetails.description || '<em>No description provided.</em>',
-      createdAt: prDetails.createdAt,
-      authorAssociation: 'OWNER',
-      isDescription: true,
-    },
-    ...comments,
-  ];
-
-  const colors = {
-    canvas: {
-      default: UI_COLORS.black,
-      subtle: UI_COLORS.surfaceElevated,
-      box: UI_COLORS.black,
-    },
-    border: {
-      default: alpha(UI_COLORS.white, 0.1),
-      muted: alpha(UI_COLORS.white, 0.08),
-    },
-    fg: {
-      default: alpha(UI_COLORS.white, 0.85),
-      muted: STATUS_COLORS.open,
-    },
-    accent: {
-      fg: STATUS_COLORS.info,
-    },
-    timeline: {
-      line: alpha(UI_COLORS.white, 0.1),
-    },
-  };
-
-  return (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 3,
-        pt: 2,
-        maxWidth: '960px',
-        mx: 'auto',
-        position: 'relative',
-      }}
-    >
-      {allItems.map((item, index) => (
-        <Box
-          key={item.id}
-          sx={{
-            display: 'flex',
-            gap: { xs: 1.25, sm: 2 },
-            position: 'relative',
-            zIndex: 1,
-            '&::before': {
-              content: index !== allItems.length - 1 ? '""' : 'none',
-              position: 'absolute',
-              top: { xs: '32px', sm: '40px' },
-              bottom: { xs: '-16px', sm: '-24px' },
-              left: { xs: '15px', sm: '20px' },
-              width: '2px',
-              backgroundColor: colors.timeline.line,
-              zIndex: 0,
-            },
-          }}
-        >
-          {/* Avatar Area */}
-          <Box sx={{ position: 'relative', zIndex: 1, flexShrink: 0 }}>
-            <Link
-              href={item.user.htmlUrl}
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <Avatar
-                src={item.user.avatarUrl}
-                alt={item.user.login}
-                sx={{
-                  width: { xs: 32, sm: 40 },
-                  height: { xs: 32, sm: 40 },
-                  border: `1px solid ${alpha(UI_COLORS.white, 0.1)}`,
-                  backgroundColor: UI_COLORS.black,
-                }}
-              />
-            </Link>
-          </Box>
-
-          {/* Comment Bubble */}
-          <Paper
-            elevation={0}
-            sx={{
-              flex: 1,
-              minWidth: 0,
-              backgroundColor: colors.canvas.box,
-              border: `1px solid ${colors.border.default}`,
-              borderRadius: '6px',
-              position: 'relative',
-              '&::after': {
-                content: { xs: 'none', sm: '""' },
-                position: 'absolute',
-                top: '11px',
-                right: '100%',
-                left: '-8px',
-                width: '16px',
-                height: '16px',
-                backgroundColor: colors.canvas.subtle, // Match the header background
-                borderBottom: `1px solid ${colors.border.default}`, // Only bottom and left create the visible arrow borders
-                borderLeft: `1px solid ${colors.border.default}`,
-                transform: 'rotate(45deg)',
-              },
-            }}
-          >
-            {/* Header */}
-            <Box
-              sx={{
-                px: { xs: 1.5, sm: 2 },
-                py: { xs: 1, sm: 1.5 }, // Slightly more vertical padding on larger screens
-                backgroundColor: colors.canvas.subtle,
-                borderBottom: `1px solid ${colors.border.default}`,
-                borderTopLeftRadius: '6px',
-                borderTopRightRadius: '6px',
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'space-between',
-                gap: 1,
-                flexWrap: 'wrap',
-                color: colors.fg.muted,
-                fontSize: { xs: '13px', sm: '14px' },
-                position: 'relative',
-                zIndex: 1, // Ensure above the arrow pseudo-element's main body
-              }}
-            >
-              <Box
-                sx={{
-                  display: 'flex',
-                  alignItems: 'center',
-                  gap: 1,
-                  flexWrap: 'wrap',
-                  minWidth: 0,
-                }}
-              >
-                <Link
-                  href={item.user.htmlUrl}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  sx={{
-                    color: colors.fg.default,
-                    fontWeight: 600,
-                    textDecoration: 'none',
-                    '&:hover': {
-                      textDecoration: 'underline',
-                      color: colors.accent.fg,
-                    },
-                  }}
-                >
-                  {item.user.login}
-                </Link>
-                <Typography
-                  component="span"
-                  sx={{ fontSize: 'inherit', color: 'inherit' }}
-                >
-                  commented
-                </Typography>
-                <Typography
-                  component="span"
-                  sx={{ fontSize: 'inherit', color: 'inherit' }}
-                >
-                  {formatDate(item.createdAt)}
-                </Typography>
-              </Box>
-
-              <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                {/* Author Association Badge */}
-                {item.authorAssociation &&
-                  item.authorAssociation !== 'NONE' && (
-                    <Chip
-                      variant="status"
-                      label={item.authorAssociation
-                        .toLowerCase()
-                        .replace('_', ' ')}
-                      sx={{
-                        color: colors.fg.muted,
-                        borderColor: colors.border.default,
-                        textTransform: 'capitalize',
-                      }}
-                    />
-                  )}
-                {/* Description Badge - Special styling */}
-                {item.isDescription && (
-                  <Chip
-                    variant="status"
-                    label="Description"
-                    sx={{
-                      color: STATUS_COLORS.info,
-                      borderColor: alpha(STATUS_COLORS.info, 0.4),
-                    }}
-                  />
-                )}
-              </Box>
-            </Box>
-
-            {/* Markdown Content */}
-            <Box
-              sx={{
-                p: { xs: 1.5, sm: 2, md: 3 }, // More padding on larger screens
-                color: colors.fg.default,
-                fontSize: { xs: '13px', sm: '14px' },
-                lineHeight: 1.6,
-                fontFamily:
-                  '-apple-system,BlinkMacSystemFont,"Segoe UI","Noto Sans",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji"', // GitHub's exact font stack
-                overflowX: 'auto',
-                ...scrollbarSx,
-                // Typography refinements
-                '& > *:first-of-type': { mt: 0 },
-                '& > *:last-child': { mb: 0 },
-                '& h1, & h2, & h3, & h4, & h5, & h6': {
-                  mt: 3,
-                  mb: 2,
-                  fontWeight: 600,
-                  lineHeight: 1.25,
-                  color: colors.fg.default,
-                  borderBottom: `1px solid ${colors.border.muted}`,
-                  pb: 0.3,
-                },
-                '& h1': { fontSize: '2em' },
-                '& h2': { fontSize: '1.5em' },
-                '& h3': { fontSize: '1.25em' },
-                '& p': { mb: 2, mt: 0 },
-                '& a': { color: colors.accent.fg, textDecoration: 'none' },
-                '& a:hover': { textDecoration: 'underline' },
-                '& ul, & ol': {
-                  pl: '2em',
-                  mb: 2,
-                  listStyleType: 'disc',
-                },
-                '& ol': { listStyleType: 'decimal' },
-                '& li': { mb: 0.5 },
-                '& li + li': { mt: '0.25em' },
-                '& blockquote': {
-                  padding: '0 1em',
-                  color: colors.fg.muted,
-                  borderLeft: `0.25em solid ${colors.border.default}`,
-                  my: 2,
-                  mx: 0,
-                },
-                '& input[type="checkbox"]': {
-                  mr: 0.5,
-                  verticalAlign: 'middle',
-                },
-                '& code': {
-                  padding: '0.2em 0.4em',
-                  margin: 0,
-                  fontSize: '85%',
-                  backgroundColor: alpha(STATUS_COLORS.neutral, 0.4),
-                  borderRadius: '6px',
-                },
-                '& pre': {
-                  mt: 2,
-                  mb: 2,
-                  p: 2,
-                  borderRadius: '6px',
-                  overflow: 'auto',
-                  backgroundColor: UI_COLORS.surfaceElevated,
-                  border: `1px solid ${colors.border.default}`,
-                  '& code': {
-                    backgroundColor: 'transparent',
-                    p: 0,
-                    fontSize: '100%',
-                  },
-                },
-                '& table': {
-                  borderCollapse: 'collapse',
-                  width: '100%',
-                  mb: 2,
-                  overflowX: 'auto',
-                },
-                '& th, & td': {
-                  border: `1px solid ${colors.border.default}`,
-                  padding: '6px 13px',
-                },
-                '& th': { fontWeight: 600 },
-                '& tr:nth-of-type(2n)': {
-                  backgroundColor: UI_COLORS.surfaceElevated,
-                },
-                '& hr': {
-                  height: '0.25em',
-                  my: 3,
-                  backgroundColor: colors.border.default,
-                  border: 0,
-                },
-                '& img': {
-                  maxWidth: '100%',
-                  borderRadius: '6px',
-                  backgroundColor: 'transparent',
-                },
-                '& .markdown-body': {
-                  backgroundColor: 'transparent',
-                  color: colors.fg.default,
-                  fontFamily: 'inherit',
-                  fontSize: '14px',
-                  lineHeight: 1.6,
-                },
-              }}
-            >
-              <div className="markdown-body" style={{ fontSize: '14px' }}>
-                <ReactMarkdown
-                  remarkPlugins={[remarkGfm]}
-                  rehypePlugins={[rehypeRaw]}
-                >
-                  {item.body}
-                </ReactMarkdown>
-              </div>
-            </Box>
-          </Paper>
-        </Box>
-      ))}
-    </Box>
-  );
+  return <ConversationTimeline items={items} />;
 };
 
 export default PRComments;
